+++ sdk/monitor/azure-monitor-opentelemetry-exporter/azure/monitor/opentelemetry/exporter/__init__.py
@@ -4,8 +4,8 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------
 
-from azure.opentelemetry.exporter.azuremonitor.export.trace._exporter import AzureMonitorTraceExporter
-from azure.opentelemetry.exporter.azuremonitor._options import ExporterOptions
+from azure.monitor.opentelemetry.exporter.export.trace._exporter import AzureMonitorTraceExporter
+from azure.monitor.opentelemetry.exporter._options import ExporterOptions
 from ._version import VERSION
 
 __all__ = ["AzureMonitorTraceExporter", "ExporterOptions"]